*  Slide TH1I under the hood of CSpectrum1DL. *  Make daqdatatypes.h use RTypes.h for e.g. UInt_t *  Resolve the fallout of Root type decisions (e.g. Size_t is a real of    some sort).

+++ main/Core/Gamma1DW.cpp
@@ -234,7 +234,7 @@ CGamma1DW::CreateStorage()
 {
   setStorageType(keWord);
   Size_t nBytes = StorageNeeded();
-  UShort_t* pStorage = new UShort_t[nBytes/sizeof(UShort_t)];
+  UShort_t* pStorage = new UShort_t[static_cast<unsigned>(nBytes/sizeof(UShort_t))];
   ReplaceStorage(pStorage);
   Clear();
   createStatArrays(1);
